move FnameTemplate declaration/implementation out of src/io/io.(cpp|h) into new files
